15-05-2015

+++ src/Funcion/Coseno.java
@@ -21,7 +21,7 @@ public class Coseno extends DecoradorOperacion{
     } 
 
     @Override
-    public Float Resultado() {
-        return (float) Math.cos(X.Resultado());
+    public Double Resultado() {
+        return Math.cos(X.Resultado());
     }
 }
