+++ src/Funcion/Resta.java
@@ -19,7 +19,7 @@ public class Resta extends DecoradorOperacion  {
     }   
     
     @Override
-    public Float Resultado() {
+    public Double Resultado() {
         return X.Resultado() - Y.Resultado();
     }
 }
